android: account links
fix: account links
refactor: remove unused
