+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/service/tunnel/TunnelManager.kt
@@ -8,11 +8,12 @@ import nym_vpn_lib.AccountStateSummary
 interface TunnelManager {
 	suspend fun stop()
 	suspend fun start(fromBackground: Boolean = true)
-	suspend fun storeMnemonic(credential: String)
+	suspend fun storeMnemonic(mnemonic: String)
 	suspend fun isMnemonicStored(): Boolean
 	suspend fun removeMnemonic()
 	suspend fun getAccountSummary(): AccountStateSummary
 	suspend fun getAccountLinks(): AccountLinks?
+	suspend fun refreshAccountLinks()
 	val stateFlow: Flow<TunnelManagerState>
 	fun getState(): Tunnel.State
 }
