+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/settings/credential/CredentialViewModel.kt
@@ -7,22 +7,17 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.launch
 import net.nymtech.nymvpn.R
-import net.nymtech.nymvpn.data.SettingsRepository
 import net.nymtech.nymvpn.service.tunnel.TunnelManager
 import net.nymtech.nymvpn.ui.common.snackbar.SnackbarController
 import net.nymtech.nymvpn.util.StringValue
-import net.nymtech.vpn.backend.Backend
 import timber.log.Timber
 import javax.inject.Inject
-import javax.inject.Provider
 
 @HiltViewModel
 class CredentialViewModel
 @Inject
 constructor(
 	private val tunnelManager: TunnelManager,
-	private val backend: Provider<Backend>,
-	private val settingsRepository: SettingsRepository,
 ) : ViewModel() {
 
 	private val _success = MutableSharedFlow<Boolean?>()
@@ -32,8 +27,6 @@ constructor(
 		runCatching {
 			tunnelManager.storeMnemonic(mnemonic.trim())
 			Timber.d("Imported account successfully")
-			val env = settingsRepository.getEnvironment()
-			backend.get().init(env, settingsRepository.isCredentialMode())
 			SnackbarController.showMessage(StringValue.StringResource(R.string.device_added_success))
 			_success.emit(true)
 		}.onFailure {
