--- nym-vpn-android/localization-util/.gitignore
@@ -1,1 +1,0 @@
-/build
