--- nym-vpn-android/localization-util/build.gradle.kts
@@ -1,54 +1,0 @@
-plugins {
-	alias(libs.plugins.android.library)
-	alias(libs.plugins.jetbrainsKotlinAndroid)
-}
-
-android {
-	namespace = "net.nymtech.localizationutil"
-	compileSdk = Constants.COMPILE_SDK
-
-	defaultConfig {
-		minSdk = Constants.MIN_SDK
-
-		testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-		consumerProguardFiles("consumer-rules.pro")
-	}
-
-	buildTypes {
-		release {
-			isMinifyEnabled = false
-			proguardFiles(
-				getDefaultProguardFile("proguard-android-optimize.txt"),
-				"proguard-rules.pro",
-			)
-		}
-		create(Constants.PRERELEASE) {
-			initWith(buildTypes.getByName(Constants.RELEASE))
-		}
-
-		create(Constants.NIGHTLY) {
-			initWith(buildTypes.getByName(Constants.RELEASE))
-		}
-	}
-	compileOptions {
-		sourceCompatibility = Constants.JAVA_VERSION
-		targetCompatibility = Constants.JAVA_VERSION
-	}
-	kotlinOptions {
-		jvmTarget = Constants.JVM_TARGET
-	}
-	buildFeatures {
-		buildConfig = true
-	}
-}
-
-dependencies {
-
-	implementation(libs.androidx.core.ktx)
-	implementation(libs.material)
-	testImplementation(libs.junit)
-	androidTestImplementation(libs.androidx.junit)
-	androidTestImplementation(libs.androidx.espresso.core)
-
-	detektPlugins(libs.detekt.rules.compose)
-}
