--- nym-vpn-android/localization-util/src/main/java/net/nymtech/localizationutil/LocaleStorage.kt
@@ -1,16 +1,0 @@
-package net.nymtech.localizationutil
-
-import android.content.Context
-import android.content.SharedPreferences
-
-class LocaleStorage(context: Context) {
-	private var preferences: SharedPreferences = context.getSharedPreferences("sp", Context.MODE_PRIVATE)
-
-	fun getPreferredLocale(): String {
-		return preferences.getString("preferred_locale", LocaleUtil.OPTION_PHONE_LANGUAGE)!!
-	}
-
-	fun setPreferredLocale(localeCode: String) {
-		preferences.edit().putString("preferred_locale", localeCode).apply()
-	}
-}
